+++ packages/mind/src/plugins/with-abstract.ts
@@ -10,15 +10,18 @@ import {
     transformPoint
 } from '@plait/core';
 import { AbstractNode, LayoutNode, MindLayoutType, isHorizontalLayout, isStandardLayout } from '@plait/layouts';
-import { AbstractHandlePosition, MindElement } from '../interfaces';
+import { AbstractHandlePosition, AbstractResizeState, MindElement, PlaitAbstractBoard } from '../interfaces';
 import { MindNodeComponent, MindQueries } from '../public-api';
-import { findLocationLeftIndex, getHitAbstractHandle, getLocationScope } from '../utils/abstract/resize';
+import { findLocationLeftIndex, getHitAbstractHandle, getLocationScope, handleTouchedAbstract } from '../utils/abstract/resize';
 import { separateChildren } from '../utils/abstract/common';
 
 export const withAbstract: PlaitPlugin = (board: PlaitBoard) => {
+    const newBoard = board as PlaitBoard & PlaitAbstractBoard;
+
     const { mousedown, mousemove, mouseup } = board;
     let activeAbstractElement: PlaitElement | undefined;
-    let abstractHandlePosition: AbstractHandlePosition | null;
+    let abstractHandlePosition: AbstractHandlePosition | undefined;
+    let touchedAbstract: PlaitElement | undefined;
     let startPoint: Point | undefined;
     let newProperty: { end: number } | { start: number } | undefined;
 
@@ -29,6 +32,9 @@ export const withAbstract: PlaitPlugin = (board: PlaitBoard) => {
 
         activeAbstractElement = activeAbstractElements.find(element => {
             abstractHandlePosition = getHitAbstractHandle(board, element as MindElement, point);
+            if (newBoard?.abstractResize) {
+                newBoard.abstractResize(AbstractResizeState.start);
+            }
             return abstractHandlePosition;
         });
 
@@ -44,6 +50,8 @@ export const withAbstract: PlaitPlugin = (board: PlaitBoard) => {
         getSelectedElements(board);
         const host = BOARD_TO_HOST.get(board);
         const endPoint = transformPoint(board, toPoint(event.x, event.y, host!));
+
+        touchedAbstract = handleTouchedAbstract(board, touchedAbstract, endPoint);
 
         if (abstractHandlePosition && activeAbstractElement) {
             const abstractComponent = PlaitElement.getComponent(activeAbstractElement) as MindNodeComponent;
@@ -64,6 +72,10 @@ export const withAbstract: PlaitPlugin = (board: PlaitBoard) => {
                 if ((activeAbstractElement as MindElement).start! >= rightNodeCount) {
                     children = leftChildren;
                 }
+            }
+
+            if (newBoard?.abstractResize) {
+                newBoard.abstractResize(AbstractResizeState.resizing);
             }
 
             const resizingLocation = isHorizontal ? endPoint[1] : endPoint[0];
@@ -93,15 +105,19 @@ export const withAbstract: PlaitPlugin = (board: PlaitBoard) => {
                     abstractHandlePosition === AbstractHandlePosition.start ? { start: locationIndex + 1 } : { end: locationIndex };
             }
 
-            abstractComponent!.updateAbstractIncludedOutline(location, abstractHandlePosition);
+            abstractComponent!.updateAbstractIncludedOutline(abstractHandlePosition, location);
         }
         mousemove(event);
     };
 
     board.mouseup = (event: MouseEvent) => {
         startPoint = undefined;
-        abstractHandlePosition = null;
+        abstractHandlePosition = undefined;
         if (activeAbstractElement) {
+            if (newBoard?.abstractResize) {
+                newBoard.abstractResize(AbstractResizeState.end);
+            }
+
             if (newProperty) {
                 const path = PlaitBoard.findPath(board, activeAbstractElement);
                 Transforms.setNode(board, newProperty, path);
@@ -109,6 +125,7 @@ export const withAbstract: PlaitPlugin = (board: PlaitBoard) => {
                 const abstractComponent = PlaitElement.getComponent(activeAbstractElement) as MindNodeComponent;
                 abstractComponent!.updateAbstractIncludedOutline();
             }
+            activeAbstractElement = undefined;
         }
         mouseup(event);
     };