+++ packages/mind/src/queries/get-branch-layouts.ts
@@ -3,7 +3,7 @@ import { findParentElement } from '../utils';
 import { MindLayoutType } from '@plait/layouts';
 import { getCorrectLayoutByElement } from './get-correct-layout-by-element';
 
-export const getBranchMindmapLayouts = (element: MindElement) => {
+export const getBranchLayouts = (element: MindElement) => {
     const layouts: MindLayoutType[] = [];
     if (element.layout) {
         //getCorrectLayoutByElement含有递归操作，getBranchMindmapLayouts本身也有递归操作，有待优化