+++ packages/mind/src/queries/index.ts
@@ -1,5 +1,5 @@
 import { getAvailableSubLayoutsByElement } from './get-available-sublayouts-by-element';
-import { getBranchMindmapLayouts } from './get-branch-mindmap-layouts-by-element';
+import { getBranchLayouts } from './get-branch-layouts';
 import { getCorrectLayoutByElement } from './get-correct-layout-by-element';
 import { getLayoutByElement } from './get-layout-by-element';
 import { getLayoutParentByElement } from './get-layout-parent-by-element';
@@ -7,7 +7,7 @@ import { getLayoutParentByElement } from './get-layout-parent-by-element';
 export const MindQueries = {
     getAvailableSubLayoutsByElement,
     getLayoutParentByElement,
-    getBranchMindmapLayouts,
+    getBranchLayouts,
     getLayoutByElement,
     getCorrectLayoutByElement
 };