+++ packages/mind/src/transforms/index.ts
@@ -1,10 +1,11 @@
 import { setLayout } from './layout';
-import { addEmoji, removeEmoji, setTopic, setTopicSize } from './node';
+import { addEmoji, removeEmoji, setTopic, setTopicSize, replaceEmoji } from './node';
 
 export const MindTransforms = {
     setLayout,
     setTopic,
     setTopicSize,
     addEmoji,
-    removeEmoji
+    removeEmoji,
+    replaceEmoji
 };