+++ packages/mind/src/transforms/layout.ts
@@ -1,8 +1,16 @@
 import { PlaitBoard, Path, PlaitNode, Transforms } from '@plait/core';
-import { isHorizontalLogicLayout, isVerticalLogicLayout, MindLayoutType } from '@plait/layouts';
+import { isHorizontalLogicLayout, isStandardLayout, isVerticalLogicLayout, MindLayoutType } from '@plait/layouts';
+import { MindElement, PlaitMind } from '../interfaces/element';
+import { handleAbstractIncluded } from '../utils/abstract/common';
 
 export const setLayout = (board: PlaitBoard, layout: MindLayoutType, path: Path) => {
     correctLogicLayoutNode(board, layout, path);
+    const element = PlaitNode.get(board, path) as MindElement;
+
+    if (PlaitMind.isMind(element) && isStandardLayout(layout)) {
+        handleAbstractIncluded(board, element);
+    }
+
     Transforms.setNode(board, { layout }, path);
 };