+++ packages/mind/src/transforms/node.ts
@@ -28,9 +28,10 @@ export const setTopicSize = (board: PlaitBoard, element: MindElement, width: num
 
 export const addEmoji = (board: PlaitBoard, element: MindElement, emojiItem: EmojiItem) => {
     const emojis = element.data.emojis || [];
-    emojis.push(emojiItem);
+    const newEmojis = [...emojis];
+    newEmojis.push(emojiItem);
     const newElement = {
-        data: { topic: element.data.topic, emojis }
+        data: { topic: element.data.topic, emojis: newEmojis }
     } as MindElement;
     const path = PlaitBoard.findPath(board, element);
     Transforms.setNode(board, newElement, path);
@@ -44,6 +45,21 @@ export const removeEmoji = (board: PlaitBoard, element: MindElement<EmojiData>, 
     if (emojis.length > 0) {
         newElement.data.emojis = emojis;
     }
+    const path = PlaitBoard.findPath(board, element);
+    Transforms.setNode(board, newElement, path);
+};
+
+export const replaceEmoji = (board: PlaitBoard, element: MindElement<EmojiData>, oldEmoji: EmojiItem, newEmoji: EmojiItem) => {
+    const newElement = {
+        data: { topic: element.data.topic }
+    } as MindElement;
+    const newEmojis = element.data.emojis.map(value => {
+        if (value === oldEmoji) {
+            return newEmoji;
+        }
+        return value;
+    });
+    newElement.data.emojis = newEmojis;
     const path = PlaitBoard.findPath(board, element);
     Transforms.setNode(board, newElement, path);
 };